# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Node.js Sign-up and Verification API
  description: Node.js and MySQL - API with email sign-up, verification, authentication and forgot password
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /requests:
    get:
      summary: Get all requests
      description: Returns a list of all requests
      operationId: getAllRequests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Request"
              example:
                - id: 1
                  employeeId: 1
                  type: "leave"
                  title: "Annual Leave"
                  description: "Vacation leave for 2 weeks"
                  status: "pending"
                  priority: "medium"
                  requestDate: "2024-06-01T12:00:00Z"
                  completionDate: null
                  approverId: 2
                  details:
                    date:
                      - start-date: "2024-07-01"
                        end-date: "2024-07-14"
                        purpose: "Vacation"
                        days: 14
                  created: "2024-06-01T12:00:00Z"
                  updated: "2024-06-01T12:00:00Z"
                - id: 2
                  employeeId: 2
                  type: "equipment"
                  title: "Laptop Request"
                  description: "Requesting a new laptop"
                  status: "approved"
                  priority: "high"
                  requestDate: "2024-06-01T12:00:00Z"
                  completionDate: "2024-06-05T12:00:00Z"
                  approverId: 1
                  details:
                    items:
                      - name: "Laptop"
                        quantity: 1
                        purpose: "Development"
                  created: "2024-06-01T12:00:00Z"
                  updated: "2024-06-05T12:00:00Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Create a new request
      description: Create a new request
      operationId: createRequest
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestRequest"
            example:
              employeeId: 1
              type: "leave"
              title: "Annual Leave"
              description: "Vacation leave for 2 weeks"
              priority: "medium"
              date:
                - start-date: "2024-07-01"
                  end-date: "2024-07-14"
                  purpose: "Vacation"
      responses:
        "200":
          description: Request created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
              example:
                id: 1
                employeeId: 1
                type: "leave"
                title: "Annual Leave"
                description: "Vacation leave for 2 weeks"
                status: "pending"
                priority: "medium"
                requestDate: "2024-06-01T12:00:00Z"
                completionDate: null
                approverId: 2
                details:
                  date:
                    - start-date: "2024-07-01"
                      end-date: "2024-07-14"
                      purpose: "Vacation"
                      days: 14
                created: "2024-06-01T12:00:00Z"
                updated: "2024-06-01T12:00:00Z"
        "400":
          description: Invalid input
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /requests/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        description: Request ID
    get:
      summary: Get request by ID
      description: Returns a single request by ID
      operationId: getRequestById
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
              example:
                id: 1
                employeeId: 1
                type: "leave"
                title: "Annual Leave"
                description: "Vacation leave for 2 weeks"
                status: "pending"
                priority: "medium"
                requestDate: "2024-06-01T12:00:00Z"
                completionDate: null
                approverId: 2
                details:
                  date:
                    - start-date: "2024-07-01"
                      end-date: "2024-07-14"
                      purpose: "Vacation"
                      days: 14
                created: "2024-06-01T12:00:00Z"
                updated: "2024-06-01T12:00:00Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    put:
      summary: Update request
      description: Update request details
      operationId: updateRequest
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestRequest"
            example:
              status: "approved"
              priority: "high"
              description: "Vacation leave for 2 weeks - approved"
      responses:
        "200":
          description: Request updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
              example:
                id: 1
                employeeId: 1
                type: "leave"
                title: "Annual Leave"
                description: "Vacation leave for 2 weeks - approved"
                status: "approved"
                priority: "high"
                requestDate: "2024-06-01T12:00:00Z"
                completionDate: "2024-06-10T12:00:00Z"
                approverId: 2
                details:
                  date:
                    - start-date: "2024-07-01"
                      end-date: "2024-07-14"
                      purpose: "Vacation"
                      days: 14
                created: "2024-06-01T12:00:00Z"
                updated: "2024-06-10T12:00:00Z"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: Delete request
      description: Delete a request
      operationId: deleteRequest
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Request deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Request deleted successfully"
              example:
                message: "Request deleted successfully"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  # Workflows API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"
  schemas:
...
    Request:
      type: object
      properties:
        id:
          type: integer
          example: 1
        type:
          type: string
          example: "Leave Request"
        status:
          type: string
          enum: [pending, approved, rejected]
          example: "pending"
        description:
          type: string
          example: "Annual leave request"
        employeeId:
          type: integer
          example: 1
        workflowId:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateRequestRequest:
      type: object
      required:
        - type
        - description
        - employeeId
        - workflowId
      properties:
        type:
          type: string
          example: "Leave Request"
        description:
          type: string
          example: "Annual leave request"
        employeeId:
          type: integer
          example: 1
        workflowId:
          type: integer
          example: 1
    UpdateRequestRequest:
      type: object
      properties:
        type:
          type: string
          example: "Leave Request"
        status:
          type: string
          enum: [pending, approved, rejected]
          example: "approved"
        description:
          type: string
          example: "Annual leave request"
        employeeId:
          type: integer
          example: 1
        workflowId:
          type: integer
          example: 1
...